perf: update semantic-release assets #80
This commit updates the list of known assets that when generated in the
remote will be committed back to the repo. This reduces the need for
developers to ensure that they have ran a local full build and generated
the correct necessary assets for a successful deployment.

On branch dsande/updateAssets
Changes to be committed:
modified:   package.json